requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Pillow==11.0.0
numpy==1.26.4